+++ src/hl7/2.1/acc.ts
@@ -1,0 +1,3 @@
+import { HL7_ACC } from "@/hl7/types/acc";
+
+export type HL7_2_1_ACC = HL7_ACC;
