+++ src/hl7/types/acc.ts
@@ -1,0 +1,5 @@
+export interface HL7_ACC {
+  acc_1: string;
+  acc_2: string;
+  acc_3: string;
+}
